convert 044_init_tests
