+++ my_project_test16742493171360845266_test_init/.gitignore
@@ -1,0 +1,4 @@
+
+target/
+dbt_packages/
+logs/
