+++ my_project_test16742493171360845266_test_init/models/example/my_second_dbt_model.sql
@@ -1,0 +1,6 @@
+
+-- Use the `ref` function to select from other models
+
+select *
+from {{ ref('my_first_dbt_model') }}
+where id = 1
